[FLINK-8992] [e2e-tests] Integrate general DataStream test job with project structure
This also includes minor cleanup of WIP code in the test job.
