+++ flink-end-to-end-tests/flink-datastream-allround-test/src/main/java/org/apache/flink/streaming/tests/artificialstate/ArtificialKeyedStateBuilder.java
@@ -16,45 +16,27 @@
  * limitations under the License.
  */
 
-package org.apache.flink.streaming.tests.general;
+package org.apache.flink.streaming.tests.artificialstate;
 
-public class Event {
+import org.apache.flink.runtime.state.FunctionInitializationContext;
 
-	private final int key;
-	private final long eventTime;
-	private final long sequenceNumber;
-	private final String payload;
+import java.io.Serializable;
 
-	public Event(int key, long eventTime, long sequenceNumber, String payload) {
-		this.key = key;
-		this.eventTime = eventTime;
-		this.sequenceNumber = sequenceNumber;
-		this.payload = payload;
+public abstract class ArtificialKeyedStateBuilder<T> implements Serializable {
+
+	private static final long serialVersionUID = -5887676929924485788L;
+
+	protected final String stateName;
+
+	public ArtificialKeyedStateBuilder(String stateName) {
+		this.stateName = stateName;
 	}
 
-	public int getKey() {
-		return key;
+	public String getStateName() {
+		return stateName;
 	}
 
-	public long getEventTime() {
-		return eventTime;
-	}
+	public abstract void artificialStateForElement(T element) throws Exception;
 
-	public long getSequenceNumber() {
-		return sequenceNumber;
-	}
-
-	public String getPayload() {
-		return payload;
-	}
-
-	@Override
-	public String toString() {
-		return "Event{" +
-			"key=" + key +
-			", eventTime=" + eventTime +
-			", sequenceNumber=" + sequenceNumber +
-			", payload='" + payload + '\'' +
-			'}';
-	}
+	public abstract void initialize(FunctionInitializationContext initializationContext);
 }
